Move the function call modifier plugin into one package
